Define AgnocastSharedMemoryAllocator trait
