+++ agnocast_heaphook/src/lib.rs
@@ -1,4 +1,3 @@
-use rlsf::Tlsf;
 use std::{
     alloc::Layout,
     ffi::{CStr, CString},
@@ -7,9 +6,13 @@ use std::{
     ptr::{self, NonNull},
     sync::{
         atomic::{AtomicBool, AtomicUsize, Ordering},
-        Mutex, OnceLock,
+        OnceLock,
     },
 };
+
+use crate::tlsf::TLSFAllocator;
+
+mod tlsf;
 
 extern "C" {
     fn initialize_agnocast(
@@ -19,10 +22,6 @@ extern "C" {
     ) -> *mut c_void;
     fn agnocast_get_borrowed_publisher_num() -> u32;
 }
-
-const POINTER_SIZE: usize = std::mem::size_of::<&usize>();
-const POINTER_ALIGN: usize = std::mem::align_of::<&usize>();
-const LAYOUT_ALIGN: usize = 1; // Minimun value that is a power of 2.
 
 // See: https://doc.rust-lang.org/src/std/sys/alloc/mod.rs.html
 #[allow(clippy::if_same_then_else)]
@@ -138,12 +137,13 @@ extern "C" fn post_fork_handler_in_child() {
     IS_FORKED_CHILD.store(true, Ordering::Relaxed);
 }
 
-const FLLEN: usize = 28; // The maximum block size is (32 << 28) - 1 = 8_589_934_591 (nearly 8GiB)
-const SLLEN: usize = 64; // The worst-case internal fragmentation is ((32 << 28) / 64 - 2) = 134_217_726 (nearly 128MiB)
-type FLBitmap = u32; // FLBitmap should contain at least FLLEN bits
-type SLBitmap = u64; // SLBitmap should contain at least SLLEN bits
-type TlsfType = Tlsf<'static, FLBitmap, SLBitmap, FLLEN, SLLEN>;
-static TLSF: OnceLock<Mutex<TlsfType>> = OnceLock::new();
+static TLSF: OnceLock<TLSFAllocator> = OnceLock::new();
+
+unsafe trait AgnocastSharedMemoryAllocator {
+    fn allocate(&self, layout: Layout) -> Option<NonNull<u8>>;
+    fn reallocate(&self, ptr: NonNull<u8>, new_layout: Layout) -> Option<NonNull<u8>>;
+    fn deallocate(&self, ptr: NonNull<u8>);
+}
 
 #[cfg(not(test))]
 fn init_tlsf() {
@@ -181,83 +181,9 @@ fn init_tlsf() {
     MEMPOOL_START.store(mempool_ptr as usize, Ordering::Relaxed);
     MEMPOOL_END.store(mempool_ptr as usize + mempool_size, Ordering::Relaxed);
 
-    let mut tlsf: TlsfType = Tlsf::new();
-    tlsf.insert_free_block(pool);
-
-    if TLSF.set(Mutex::new(tlsf)).is_err() {
+    if TLSF.set(TLSFAllocator::new(pool)).is_err() {
         panic!("[ERROR] [Agnocast] TLSF is already initialized.");
     }
-}
-
-fn tlsf_allocate_wrapped(layout: Layout) -> Option<NonNull<u8>> {
-    // `alignment` must be at least `POINTER_ALIGN` to ensure that `aligned_ptr` is properly aligned to store a pointer.
-    let alignment = layout.align().max(POINTER_ALIGN);
-    let size = layout.size();
-    let layout = Layout::from_size_align(POINTER_SIZE + size + alignment, LAYOUT_ALIGN).ok()?;
-
-    // the original pointer returned by the internal allocator
-    let mut tlsf = TLSF.get().unwrap().lock().unwrap();
-    let original_ptr = tlsf.allocate(layout)?;
-    let original_addr = original_ptr.as_ptr() as usize;
-
-    // the aligned pointer returned to the user
-    //
-    // It is our responsibility to satisfy alignment constraints.
-    // We avoid using `Layout::align` because doing so requires us to remember the alignment.
-    // This is because `Tlsf::{reallocate, deallocate}` functions require the same alignment.
-    // See: https://docs.rs/rlsf/latest/rlsf/struct.Tlsf.html
-    let aligned_addr = (original_addr + POINTER_SIZE + alignment - 1) & !(alignment - 1);
-
-    // SAFETY: `aligned_addr` must be non-zero.
-    debug_assert!(aligned_addr % alignment == 0 && aligned_addr != 0);
-    let aligned_ptr = unsafe { NonNull::new_unchecked(aligned_addr as *mut u8) };
-
-    // store the original pointer
-    unsafe { *aligned_ptr.as_ptr().byte_sub(POINTER_SIZE).cast() = original_ptr };
-
-    Some(aligned_ptr)
-}
-
-fn tlsf_reallocate_wrapped(ptr: NonNull<u8>, new_layout: Layout) -> Option<NonNull<u8>> {
-    // `alignment` must be at least `POINTER_ALIGN` to ensure that `aligned_ptr` is properly aligned to store a pointer.
-    let alignment = new_layout.align().max(POINTER_ALIGN);
-    let size = new_layout.size();
-    let new_layout = Layout::from_size_align(POINTER_SIZE + size + alignment, LAYOUT_ALIGN).ok()?;
-
-    // get the original pointer
-    // SAFETY: `ptr` must have been allocated by `tlsf_allocate_wrapped`.
-    let old_original_ptr = unsafe { *ptr.as_ptr().byte_sub(POINTER_SIZE).cast() };
-
-    // the original pointer returned by the internal allocator
-    let mut tlsf = TLSF.get().unwrap().lock().unwrap();
-    let new_original_ptr = unsafe { tlsf.reallocate(old_original_ptr, new_layout) }?;
-    let new_original_addr = new_original_ptr.as_ptr() as usize;
-
-    // the aligned pointer returned to the user
-    //
-    // It is our responsibility to satisfy alignment constraints.
-    // We avoid using `Layout::align` because doing so requires us to remember the alignment.
-    // This is because `Tlsf::{reallocate, deallocate}` functions require the same alignment.
-    // See: https://docs.rs/rlsf/latest/rlsf/struct.Tlsf.html
-    let new_aligned_addr = (new_original_addr + POINTER_SIZE + alignment - 1) & !(alignment - 1);
-
-    // SAFETY: `new_aligned_addr` must be non-zero.
-    debug_assert!(new_aligned_addr % alignment == 0 && new_aligned_addr != 0);
-    let new_aligned_ptr = unsafe { NonNull::new_unchecked(new_aligned_addr as *mut u8) };
-
-    // store the original pointer
-    unsafe { *new_aligned_ptr.as_ptr().byte_sub(POINTER_SIZE).cast() = new_original_ptr };
-
-    Some(new_aligned_ptr)
-}
-
-fn tlsf_deallocate_wrapped(ptr: NonNull<u8>) {
-    // get the original pointer
-    // SAFETY: `ptr` must have been allocated by `tlsf_{allocate, reallocate}_wrapped`.
-    let original_ptr = unsafe { *ptr.as_ptr().byte_sub(POINTER_SIZE).cast() };
-
-    let mut tlsf = TLSF.get().unwrap().lock().unwrap();
-    unsafe { tlsf.deallocate(original_ptr, LAYOUT_ALIGN) }
 }
 
 #[cfg(not(test))]
@@ -307,7 +233,7 @@ pub extern "C" fn malloc(size: usize) -> *mut c_void {
         Err(_) => return ptr::null_mut(),
     };
 
-    match tlsf_allocate_wrapped(layout) {
+    match TLSF.get().unwrap().allocate(layout) {
         Some(non_null_ptr) => non_null_ptr.as_ptr().cast(),
         None => ptr::null_mut(),
     }
@@ -334,7 +260,7 @@ pub unsafe extern "C" fn free(ptr: *mut c_void) {
 
     match (is_shared, is_forked_child) {
         (true, true) => (), // In the child processes, ignore the free operation to the shared memory
-        (true, false) => tlsf_deallocate_wrapped(non_null_ptr),
+        (true, false) => TLSF.get().unwrap().deallocate(non_null_ptr),
         (false, _) => (*ORIGINAL_FREE.get_or_init(init_original_free))(ptr),
     }
 }
@@ -353,7 +279,7 @@ pub extern "C" fn calloc(num: usize, size: usize) -> *mut c_void {
         Err(_) => return ptr::null_mut(),
     };
 
-    match tlsf_allocate_wrapped(layout) {
+    match TLSF.get().unwrap().allocate(layout) {
         Some(non_null_ptr) => {
             let ptr = non_null_ptr.as_ptr();
             unsafe {
@@ -389,18 +315,18 @@ pub unsafe extern "C" fn realloc(ptr: *mut c_void, new_size: usize) -> *mut c_vo
                 Some(non_null_ptr) => {
                     // If `new_size` is equal to zero, and `ptr` is not NULL, then the call is equivalent to `free(ptr)`.
                     if new_layout.size() == 0 {
-                        tlsf_deallocate_wrapped(non_null_ptr);
+                        TLSF.get().unwrap().deallocate(non_null_ptr);
                         return ptr::null_mut();
                     }
 
-                    match tlsf_reallocate_wrapped(non_null_ptr, new_layout) {
+                    match TLSF.get().unwrap().reallocate(non_null_ptr, new_layout) {
                         Some(non_null_ptr) => non_null_ptr.as_ptr().cast(),
                         None => ptr::null_mut(),
                     }
                 }
                 None => {
                     // If `ptr` is NULL, then the call is equivalent to `malloc(size)`.
-                    match tlsf_allocate_wrapped(new_layout) {
+                    match TLSF.get().unwrap().allocate(new_layout) {
                         Some(non_null_ptr) => non_null_ptr.as_ptr().cast(),
                         None => ptr::null_mut(),
                     }
@@ -431,7 +357,7 @@ pub extern "C" fn posix_memalign(memptr: &mut *mut c_void, alignment: usize, siz
         Err(_) => return libc::ENOMEM,
     };
 
-    match tlsf_allocate_wrapped(layout) {
+    match TLSF.get().unwrap().allocate(layout) {
         Some(non_null_ptr) => {
             *memptr = non_null_ptr.as_ptr().cast();
             0
@@ -458,7 +384,7 @@ pub extern "C" fn aligned_alloc(alignment: usize, size: usize) -> *mut c_void {
         Err(_) => return ptr::null_mut(),
     };
 
-    match tlsf_allocate_wrapped(layout) {
+    match TLSF.get().unwrap().allocate(layout) {
         Some(non_null_ptr) => non_null_ptr.as_ptr().cast(),
         None => std::ptr::null_mut(),
     }
@@ -478,7 +404,7 @@ pub extern "C" fn memalign(alignment: usize, size: usize) -> *mut c_void {
         Err(_) => return ptr::null_mut(),
     };
 
-    match tlsf_allocate_wrapped(layout) {
+    match TLSF.get().unwrap().allocate(layout) {
         Some(non_null_ptr) => non_null_ptr.as_ptr().cast(),
         None => std::ptr::null_mut(),
     }
@@ -536,10 +462,8 @@ fn init_tlsf() {
     MEMPOOL_START.store(mmap_ptr as usize, Ordering::Relaxed);
     MEMPOOL_END.store(mmap_ptr as usize + mempool_size, Ordering::Relaxed);
 
-    let mut tlsf: TlsfType = Tlsf::new();
-    tlsf.insert_free_block(pool);
-
-    TLSF.set(Mutex::new(tlsf));
+    // FIXME
+    TLSF.set(TLSFAllocator::new(pool));
 }
 
 #[cfg(test)]
